refactor(rust): extract usage.rs with analytics and limits commands
